Move camera around instead of moving the world around
before we would apply the camera transform to the objects. This made the
collision detection heavier since for each object we had to apply it.
With the new logic we move the initial rays around instead. This removes
the cost of a matrix multiplication for each object we test a collision
against.